realtime sensor

+++ lib/utils/sensor_model.dart
@@ -8,8 +8,8 @@ class SensorModel{
 });
   bool? makan = false;
   int? sensorFlow = 0;
-  double? sensorFlowTotal = 0.0;
-  int? sensorSuhu = 0;
+  double? sensorFlowTotal = 0;
+  double? sensorSuhu = 0.0 ;
   String? sensorTinggi = '';
 
   factory SensorModel.fromJson(Map<String, dynamic> json){
@@ -21,21 +21,5 @@ class SensorModel{
       sensorTinggi: json["sensorTinggi"]
     );
   }
-  // Future<void> fetchData() async {
-  //   final url = Uri.parse('https://sakanayoi-ikan-default-rtdb.asia-southeast1.firebasedatabase.app/.json');
-  //   final response = await http.get(url);
-  //
-  //   if (response.statusCode == 200) {
-  //     final data = jsonDecode(response.body);
-  //     setState(() {
-  //       makan = data['makan'] ?? false;
-  //       sensorFlow = data['sensorFlow'] ?? 0;
-  //       sensorFlowTotal = data['sensorFlowTotal'] ?? 0.0;
-  //       sensorSuhu = data['sensorSuhu'] ?? 0.0;
-  //       sensorTinggi = data['sensorTinggi'] ?? '';
-  //     });
-  //   } else {
-  //     print('Failed to fetch data from the server');
-  //   }
-  // }
+
 }
